% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\name{Clip}
\alias{Clip}
\title{Clip a vector to lower and higher bound, setting more extreme values to bound}
\usage{
Clip(v, lower = -Inf, upper = Inf)
}
\arguments{
\item{v}{numeric array}
}
\value{
array of same length
}
\description{
Takes a numeric vector and limits its range to be within two bounds.  values are maintained but set to bound.
}
\examples{
Clip(iris$Sepal.Length, 5, 6)
}
