% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_at_sd.r
\name{ClipAtSd}
\alias{ClipAtSd}
\title{Clip a vector at certain number of standard deviations from the mean, setting more extreme values to bound}
\usage{
ClipAtSd(v, clip.at.sd = 3)
}
\arguments{
\item{v}{numeric array}
}
\value{
array of same length
}
\description{
Takes a numeric vector and limits its range to be within certain number of standard deviations.
}
\examples{
ClipAtSd(iris$Sepal.Length)
}
