% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages2.R
\name{install.packages2}
\alias{install.packages2}
\title{Install a package if necessary and load}
\usage{
install.packages2(pckg.names, also.load = TRUE)
}
\arguments{
\item{pckg.names}{string or vector of strings containing name(s) of the packages}

\item{also.load}{boolean on whether to load the package}
}
\description{
Install a package if necessary and load
}
\examples{
install.packages2(c('tidyverse', 'mclust'))
}
