% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_histogram.R
\name{pretty_histogram}
\alias{pretty_histogram}
\title{Generate a subjectively pretty ggplot histogram}
\usage{
pretty_histogram(v, ...)
}
\arguments{
\item{v}{array with data that goes into histogram}

\item{...}{any additional argument passed on to \code{\link{ggplot2::labs}} such as title, xlab, ylab.}
}
\value{
ggplot object
}
\description{
Uses a set of defaults that I like to generate a histogram
}
\examples{
pretty_histogram(iris$Sepal.Length)
}
\seealso{
\code{\link{ggplot2::geom_histogram}} which this function wraps
}
