% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_histogram.r
\name{PrettyHistogram}
\alias{PrettyHistogram}
\title{Generate a subjectively pretty ggplot histogram}
\usage{
PrettyHistogram(v, force.factor = FALSE, sort.factor = TRUE, ...)
}
\arguments{
\item{v}{array with data that goes into histogram}

\item{force.factor}{logical, convert input to factor so you ensure a percentage for each level}

\item{sort.factor}{should factor be reordered so it goes from most to least common?}

\item{...}{any additional argument passed on to \code{\link{ggplot2::labs}} such as title, xlab, ylab.}
}
\value{
ggplot object
}
\description{
Uses a set of defaults that I like to generate a histogram of either a numeric or factor
}
\examples{
PrettyHistogram(iris$Sepal.Length)
PrettyHistogram(iris$Species)
}
\seealso{
\code{\link{ggplot2::geom_histogram}} which this function wraps
}
