% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{write_xclip}
\alias{write_xclip}
\title{Write object to clipboard (e.g. dataframe)}
\usage{
write_xclip(x, ...)
}
\arguments{
\item{x}{the object to copy to clipboard, usually a dataframe}

\item{...}{any additional argument passed on to \code{\link{write.table}}}
}
\value{
TRUE if successful, FALSE if xclip is not available on your system.
}
\description{
Write a dataframe or any other object to your clipboard, ready to paste into a spreadsheet or presentation. Not sure if this works anywhere other than Linux.
}
\examples{
write_xclip(iris)
}
\seealso{
\code{\link{write.table}} which this function wraps
}
