% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor_histogram.R
\name{plot_factor_histogram}
\alias{plot_factor_histogram}
\title{Plot a histogram with factor frequencies}
\usage{
plot_factor_histogram(v, max.levels = 20)
}
\arguments{
\item{v}{factor or other array to plot}

\item{max.levels}{if factor has more than this number of levels, return empty ggplot object}
}
\value{
ggplot object
}
\description{
Pass a factor and get a ggplot histogram of the frequency of each factor level. If no factor is passed, the data are converted to factor.
If the factor has more than \code{max.levels} levels, abandon ship and return empty ggplot object.
}
\examples{
plot_factor_histogram(iris)
}
\seealso{
\code{\link{geom_bar}} which this function wraps
}
